Update components.tsx
To ensure that the Google Analytics event is sent before the user is redirected, I modified the handleClick function to wait for the ReactGA.outboundLink function to complete. This function must be promisified as it does not return a promise by default.

handleAnalyticsEvent function will now returns a promise, the handleClick function can now await the analytics call. This ensures that the Google Analytics tracking is completed before the page is redirected. Note that handleClick is now an async function.