Update components.tsx
Deprecated attrs: The use of .attrs with styled-components is not recommended as it can lead to unexpected bugs and performance issues. It's better to pass props directly to styled components and interpolate functions to handle dynamic styling.

n the above code, the backgroundColor is no longer passed as an attribute but is determined directly within the styled component's template literal. The warning prop is destructured along with theme from the props to decide the color of the button. This method is cleaner and prevents the potential bugs and performance issues associated with .attrs.

Please note that you need to ensure your theme provider is correctly set up to provide theme.red1, theme.primary1, theme.white, theme.bg1, and theme.text4. Also, make sure to replace all instances where Button is used with the correct prop warning.